backend/requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# AI/ML - IBM Granite Model
transformers==4.37.2
torch==2.2.0
accelerate==0.26.1
sentencepiece==0.1.99

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
pyarrow==15.0.0

# Google Cloud TTS
google-cloud-texttospeech==2.16.3

# Text Processing
regex==2023.12.25
Unidecode==1.3.8

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Frontend
streamlit==1.31.0
altair==5.2.0
plotly==5.18.0
